refactor(grid): move PiecewiseConstantFn and PiecewiseLinearFn to treetime-grid
- Move piecewise function primitives from coalescent/ to treetime-grid crate
- These are general-purpose math primitives (non-uniform breakpoint counterparts to GridFn), not coalescent-specific
- Promote treetime-grid from dev-dependency to regular dependency in treetime crate
- Carry over all tests with updated imports
- Rewrite doc comments to be domain-neutral with thorough mathematical descriptions
