+++ packages/treetime-grid/src/__tests__/mod.rs
@@ -1,2 +1,4 @@
 mod grid;
 mod grid_fn;
+mod piecewise_constant_fn;
+mod piecewise_linear_fn;
