+++ packages/treetime/Cargo.toml
@@ -17,6 +17,7 @@ clap = ["dep:clap"]
 # [internal]
 treetime-distribution = { workspace = true }
 treetime-graph = { workspace = true }
+treetime-grid = { workspace = true }
 treetime-io = { workspace = true }
 treetime-primitives = { workspace = true }
 treetime-utils = { workspace = true }
@@ -77,7 +78,6 @@ ctor = { workspace = true }
 indoc = { workspace = true }
 lazy_static = { workspace = true }
 petgraph = { workspace = true }
-treetime-grid = { workspace = true }
 pretty_assertions = { workspace = true }
 proptest = { workspace = true }
 rayon = { workspace = true }
