+++ packages/treetime/src/coalescent/__tests__/mod.rs
@@ -5,7 +5,5 @@ mod test_gm_total_lh;
 mod test_integration;
 mod test_lineage_dynamics;
 mod test_optimize_tc;
-mod test_piecewise_constant_fn;
-mod test_piecewise_linear_fn;
 mod test_skyline;
 mod test_total_lh;
