+++ packages/treetime/src/coalescent/integration.rs
@@ -1,8 +1,8 @@
-use crate::coalescent::piecewise_constant_fn::PiecewiseConstantFn;
-use crate::coalescent::piecewise_linear_fn::PiecewiseLinearFn;
 use eyre::Report;
 use ndarray::Array1;
 use treetime_distribution::Distribution;
+use treetime_grid::piecewise_constant_fn::PiecewiseConstantFn;
+use treetime_grid::piecewise_linear_fn::PiecewiseLinearFn;
 use treetime_utils::make_error;
 
 /// Computes branch merger rate κ(t) and total merger rate λ(t).
