+++ packages/treetime/src/coalescent/mod.rs
@@ -5,8 +5,6 @@ mod events;
 mod integration;
 mod lineage_dynamics;
 pub mod optimize_tc;
-mod piecewise_constant_fn;
-mod piecewise_linear_fn;
 pub mod skyline;
 mod time_coordinate;
 pub mod total_lh;
